refactor(cli): breaking up acc-packet logic within its own mod
